+++ test/unit/mongo_rep_admin/chk_rep_lag.py
@@ -22,47 +22,10 @@ import mock
 # Local
 sys.path.append(os.getcwd())
 import mongo_rep_admin                          # pylint:disable=E0401,C0413
+import lib.gen_class as gen_class           # pylint:disable=E0401,C0413,R0402
 import version                                  # pylint:disable=E0401,C0413
 
 __version__ = version.__version__
-
-
-class ArgParser():                                      # pylint:disable=R0903
-
-    """Class:  ArgParser
-
-    Description:  Class stub holder for gen_class.ArgParser class.
-
-    Methods:
-        __init__
-        get_val
-
-    """
-
-    def __init__(self):
-
-        """Method:  __init__
-
-        Description:  Class initialization.
-
-        Arguments:
-
-        """
-
-        self.cmdline = None
-        self.args_array = {}
-
-    def get_val(self, skey, def_val=None):
-
-        """Method:  get_val
-
-        Description:  Method stub holder for gen_class.ArgParser.get_val.
-
-        Arguments:
-
-        """
-
-        return self.args_array.get(skey, def_val)
 
 
 class Server():                                         # pylint:disable=R0903
@@ -114,9 +77,6 @@ class UnitTest(unittest.TestCase):
 
     Methods:
         setUp
-        test_mongo_json_fail
-        test_mongo_std_fail
-        test_mongo_successful
         test_secondary_chk
         test_primary_chk
 
@@ -134,8 +94,8 @@ class UnitTest(unittest.TestCase):
 
         self.get_master = {"optimeDate": "2019-07-26 11:13:02"}
         self.server = Server()
-        self.args = ArgParser()
-        self.args.args_array = {"-m": "config", "-d": "directory/path"}
+        self.data_config = {"to_addr": "ToAddress", "suppress": True}
+        self.dtg = gen_class.TimeFormat()
         self.status = (True, None)
         self.status2 = (False, "_process_std: Error Message")
         self.status3 = (False, "_process_json: Error Message")
@@ -143,68 +103,6 @@ class UnitTest(unittest.TestCase):
         self.results2 = (False, "_process_std: Error Message")
         self.results3 = (False, "_process_json: Error Message")
 
-    @mock.patch("mongo_rep_admin.gen_libs.load_module",
-                mock.Mock(return_value="Mongo_Cfg"))
-    @mock.patch("mongo_rep_admin.chk_mem_rep_lag")
-    @mock.patch("mongo_rep_admin.get_master")
-    def test_mongo_json_fail(self, mock_mst, mock_lag):
-
-        """Function:  test_mongo_json_fail
-
-        Description:  Test with failed connection to Mongo for json.
-
-        Arguments:
-
-        """
-
-        mock_mst.return_value = self.get_master
-        mock_lag.return_value = self.status3
-
-        self.assertEqual(
-            mongo_rep_admin.chk_rep_lag(self.server, self.args), self.results3)
-
-    @mock.patch("mongo_rep_admin.gen_libs.load_module",
-                mock.Mock(return_value="Mongo_Cfg"))
-    @mock.patch("mongo_rep_admin.chk_mem_rep_lag")
-    @mock.patch("mongo_rep_admin.get_master")
-    def test_mongo_std_fail(self, mock_mst, mock_lag):
-
-        """Function:  test_mongo_std_fail
-
-        Description:  Test with failed connection to Mongo for std.
-
-        Arguments:
-
-        """
-
-        mock_mst.return_value = self.get_master
-        mock_lag.return_value = self.status2
-
-        self.assertEqual(
-            mongo_rep_admin.chk_rep_lag(self.server, self.args), self.results2)
-
-    @mock.patch("mongo_rep_admin.gen_libs.load_module",
-                mock.Mock(return_value="Mongo_Cfg"))
-    @mock.patch("mongo_rep_admin.chk_mem_rep_lag")
-    @mock.patch("mongo_rep_admin.get_master")
-    def test_mongo_successful(self, mock_mst, mock_lag):
-
-        """Function:  test_mongo_successful
-
-        Description:  Test with successful connection.
-
-        Arguments:
-
-        """
-
-        mock_mst.return_value = self.get_master
-        mock_lag.return_value = self.status
-
-        self.assertEqual(
-            mongo_rep_admin.chk_rep_lag(self.server, self.args), self.results)
-
-    @mock.patch("mongo_rep_admin.gen_libs.load_module",
-                mock.Mock(return_value="Mongo_Cfg"))
     @mock.patch("mongo_rep_admin.get_optimedate")
     @mock.patch("mongo_rep_admin.chk_mem_rep_lag")
     @mock.patch("mongo_rep_admin.get_master")
@@ -223,10 +121,9 @@ class UnitTest(unittest.TestCase):
         mock_time.return_value = "2019-07-26 11:13:02"
 
         self.assertEqual(
-            mongo_rep_admin.chk_rep_lag(self.server, self.args), self.results)
+            mongo_rep_admin.chk_rep_lag(
+                self.server, self.dtg, **self.data_config), self.results)
 
-    @mock.patch("mongo_rep_admin.gen_libs.load_module",
-                mock.Mock(return_value="Mongo_Cfg"))
     @mock.patch("mongo_rep_admin.chk_mem_rep_lag")
     @mock.patch("mongo_rep_admin.get_master")
     def test_primary_chk(self, mock_mst, mock_lag):
@@ -243,7 +140,8 @@ class UnitTest(unittest.TestCase):
         mock_lag.return_value = self.status
 
         self.assertEqual(
-            mongo_rep_admin.chk_rep_lag(self.server, self.args), self.results)
+            mongo_rep_admin.chk_rep_lag(
+                self.server, self.dtg, **self.data_config), self.results)
 
 
 if __name__ == "__main__":